Move all internal Go packages into ‘internal’ subdirectory
See https://go.dev/doc/modules/layout#package-or-command-with-supporting-packages
and https://pkg.go.dev/cmd/go#hdr-Internal_packages.
