+++ internal/docs/stardoc_output_go_proto/dummy.go
