+++ private/package_features.bzl
@@ -32,7 +32,8 @@ visibility([
     "//gazelle",
     "//gazelle/elisp",
     "//gazelle/testdata",
-    "//private/testutil",
+    "//internal/docs",
+    "//internal/testutil",
     "//tests",
     "//tests/integration",
     "//tests/integration/pkg",
